requirements.txt
chromadb==1.5.2
open-clip-torch==3.3.0
Pillow==12.0.0
streamlit==1.55.0
matplotlib==3.10.8
pandas==2.3.3
numpy==2.4.2
deep-translator==1.11.4
torch==2.10.0
torchvision==0.25.0
